% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.R
\name{.find_air}
\alias{.find_air}
\title{Locate the air binary}
\usage{
.find_air()
}
\value{
Path to the air binary.
}
\description{
Searches for \code{air} on the system PATH. If not found, tries common
installation locations. If still not found, raises an informative error
suggesting to call \code{\link[=install_air]{install_air()}}.
}
\keyword{internal}
