% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_sarif}
\alias{export_sarif}
\title{Export Results in SARIF Format}
\usage{
export_sarif(x, output = "rsonar.sarif")
}
\arguments{
\item{x}{An \code{rsonar_result} object returned by \code{\link[=sonar_analyse]{sonar_analyse()}}.}

\item{output}{Path to the output SARIF file. Default \code{"rsonar.sarif"}.}
}
\value{
The path to the generated SARIF file (invisibly).
}
\description{
Generates a \href{https://sarifweb.azurewebsites.net/}{SARIF} (Static Analysis
Results Interchange Format) file. SARIF is supported by GitHub Code Scanning,
VS Code, Azure DevOps and many other tools.
}
\examples{
\dontrun{
res <- sonar_analyse(".")
export_sarif(res, "results.sarif")
}

}
