% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debt.R
\name{print.rsonar_debt}
\alias{print.rsonar_debt}
\title{Print an rsonar_debt Object}
\usage{
\method{print}{rsonar_debt}(x, ...)
}
\arguments{
\item{x}{An \code{rsonar_debt} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
\code{x} invisibly.
}
\description{
Print an rsonar_debt Object
}
