% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{print.rsonar_diff}
\alias{print.rsonar_diff}
\title{Print an rsonar_diff Object}
\usage{
\method{print}{rsonar_diff}(x, ...)
}
\arguments{
\item{x}{An \code{rsonar_diff} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
\code{x} invisibly.
}
\description{
Print an rsonar_diff Object
}
