% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.R
\name{print.rsonar_fix}
\alias{print.rsonar_fix}
\title{Print an rsonar_fix Object}
\usage{
\method{print}{rsonar_fix}(x, ...)
}
\arguments{
\item{x}{An \code{rsonar_fix} object returned by \code{\link[=sonar_fix]{sonar_fix()}}.}

\item{...}{Additional arguments (ignored).}
}
\value{
\code{x} invisibly.
}
\description{
Print an rsonar_fix Object
}
