% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_gate.R
\name{print.rsonar_gate}
\alias{print.rsonar_gate}
\title{Print an rsonar_gate Object}
\usage{
\method{print}{rsonar_gate}(x, ...)
}
\arguments{
\item{x}{An \code{rsonar_gate} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
\code{x} invisibly.
}
\description{
Print an rsonar_gate Object
}
