% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspots.R
\name{print.rsonar_hotspots}
\alias{print.rsonar_hotspots}
\title{Print an rsonar_hotspots Object}
\usage{
\method{print}{rsonar_hotspots}(x, ...)
}
\arguments{
\item{x}{An \code{rsonar_hotspots} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
\code{x} invisibly.
}
\description{
Print an rsonar_hotspots Object
}
