% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.rsonar_result}
\alias{print.rsonar_result}
\title{Print an rsonar_result Object}
\usage{
\method{print}{rsonar_result}(x, ...)
}
\arguments{
\item{x}{An \code{rsonar_result} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
\code{x} invisibly.
}
\description{
Print an rsonar_result Object
}
