% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{print.rsonar_score}
\alias{print.rsonar_score}
\title{Print an rsonar_score Object}
\usage{
\method{print}{rsonar_score}(x, ...)
}
\arguments{
\item{x}{An \code{rsonar_score} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
\code{x} invisibly.
}
\description{
Print an rsonar_score Object
}
