% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.R
\name{print.sonar_fix}
\alias{print.sonar_fix}
\title{Print a sonar_fix Object}
\usage{
\method{print}{sonar_fix}(x, ...)
}
\arguments{
\item{x}{A \code{sonar_fix} object returned by \code{\link[=sonar_fix]{sonar_fix()}}.}

\item{...}{Additional arguments (ignored).}
}
\value{
\code{x} invisibly.
}
\description{
Print a sonar_fix Object
}
