% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.rsonar_result}
\alias{summary.rsonar_result}
\title{Summarise an rsonar_result Object}
\usage{
\method{summary}{rsonar_result}(object, ...)
}
\arguments{
\item{object}{An \code{rsonar_result} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
\code{object} invisibly.
}
\description{
Summarise an rsonar_result Object
}
