% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{use_rsonar_lintr}
\alias{use_rsonar_lintr}
\title{Copy a Reference .lintr File into a Project}
\usage{
use_rsonar_lintr(path = ".", overwrite = FALSE)
}
\arguments{
\item{path}{Target directory. Default: current directory.}

\item{overwrite}{Overwrite if a \code{.lintr} file already exists. Default \code{FALSE}.}
}
\value{
The path to the created \code{.lintr} file (invisibly).
}
\description{
Copies the default \code{.lintr} configuration file recommended by
\code{rsonar} into the target directory. This file enables linters equivalent
to the most common SonarQube rules for R.
}
\examples{
\dontrun{
use_rsonar_lintr()
}

}
